adaptive & responsive update and refactor and update download button link
