+++ lib/view/intro/components/download_button.dart
@@ -10,17 +10,19 @@ class DownloadButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return InkWell(
       onTap: () {
-        launchUrl(Uri.parse('https://drive.google.com/file/d/1HSIe7rdk8VtrAL4DQuybfMHQgDrQ6xNs/view?usp=sharing'));
+        launchUrl(Uri.parse(
+            'https://drive.google.com/file/d/1uKkGWhZxPWlzuk3jp4wxqre-NBMqZ5Pl/view'));
       },
       child: Container(
         alignment: Alignment.center,
-        padding: const EdgeInsets.symmetric(vertical: defaultPadding/1.5,horizontal: defaultPadding*2),
+        padding: const EdgeInsets.symmetric(
+            vertical: defaultPadding / 1.5, horizontal: defaultPadding * 2),
         decoration: BoxDecoration(
           borderRadius: BorderRadius.circular(20),
-            boxShadow:const [
-              BoxShadow(color: Colors.blue,offset: Offset(0, -1),blurRadius: 5),
-              BoxShadow(color: Colors.red,offset: Offset(0, 1),blurRadius: 5),
-            ],
+          boxShadow: const [
+            BoxShadow(color: Colors.blue, offset: Offset(0, -1), blurRadius: 5),
+            BoxShadow(color: Colors.red, offset: Offset(0, 1), blurRadius: 5),
+          ],
           gradient: LinearGradient(
               begin: Alignment.topLeft,
               end: Alignment.bottomRight,
@@ -38,9 +40,14 @@ class DownloadButton extends StatelessWidget {
                   letterSpacing: 1.2,
                   fontWeight: FontWeight.bold),
             ),
-            const SizedBox(width: defaultPadding/3,),
-            const Icon(FontAwesomeIcons.download,color: Colors.white70,size: 15,)
-
+            const SizedBox(
+              width: defaultPadding / 3,
+            ),
+            const Icon(
+              FontAwesomeIcons.download,
+              color: Colors.white70,
+              size: 15,
+            )
           ],
         ),
       ),
