+++ lib/view/projects/components/project_media_viewer.dart
@@ -1,11 +1,8 @@
 import 'dart:developer';
 
 import 'package:flutter/material.dart';
-import 'package:video_player/video_player.dart';
-import 'package:chewie/chewie.dart';
-import 'package:photo_view/photo_view.dart';
-import 'animated_gradient_image_container.dart';
-import 'hoverable_image_container.dart';
+import 'media_viewer_desktop.dart';
+import 'media_viewer_mobile.dart';
 
 class ProjectMediaViewer extends StatelessWidget {
   final String name;
@@ -32,19 +29,18 @@ class ProjectMediaViewer extends StatelessWidget {
     'https://samplelib.com/mp4/sample-5s.mp4'
   ];
 
-  const ProjectMediaViewer(
-      {Key? key,
-      required this.name,
-      required this.description,
-      List<String>? images,
-      List<String>? videos})
-      : images = images ?? testImages,
+  const ProjectMediaViewer({
+    Key? key,
+    required this.name,
+    required this.description,
+    List<String>? images,
+    List<String>? videos,
+  })  : images = images ?? testImages,
         videos = videos ?? testVideos,
         super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final isMobile = MediaQuery.of(context).size.width < 800;
     return Scaffold(
       appBar: AppBar(
         surfaceTintColor: Colors.transparent,
@@ -62,11 +58,12 @@ class ProjectMediaViewer extends StatelessWidget {
         child: LayoutBuilder(
           builder: (context, constraints) {
             final width = constraints.maxWidth;
-            log(width.toString());
-            final isDesktop = width >= 1000;
+            final isBigDesktop = width >= 1300;
+            final isDesktop = width >= 1000 && width < 1300;
             final isTablet = width >= 700 && width < 1000;
             final isBigMobile = width >= 600 && width < 700;
-            final isMobile = width < 700;
+            log('width $width');
+
             final gridCrossAxisCount = isDesktop
                 ? 3
                 : (isTablet
@@ -76,275 +73,32 @@ class ProjectMediaViewer extends StatelessWidget {
                         : 5);
             return CustomScrollView(
               slivers: [
-                SliverFillRemaining(
-                  hasScrollBody: false,
-                  child: Padding(
-                    padding: EdgeInsets.symmetric(
-                        horizontal: isBigMobile ? 48 : 24,
-                        vertical: isBigMobile ? 36 : 16),
-                    child: (isDesktop || isTablet)
-                        ? SizedBox(
-                            height: 400,
-                            child: Row(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                Expanded(
-                                  flex: 3,
-                                  child: VideoGalleryWithGradientContainer(
-                                      testVideos: testVideos),
-                                ),
-                                const SizedBox(width: 32),
-                                Expanded(
-                                  flex: 1,
-                                  child: ImageGallery(
-                                    images: ProjectMediaViewer.testImages,
-                                    crossAxisCount: isTablet ? 2 : 3,
-                                  ),
-                                ),
-                              ],
-                            ),
-                          )
-                        : Column(
-                            children: [
-                              Text(
-                                description,
-                                style: const TextStyle(
-                                    fontSize: 18, color: Colors.white70),
-                                textAlign: TextAlign.center,
-                              ),
-                              const SizedBox(height: 32),
-                              SizedBox(
-                                height: 220,
-                                child: VideoGalleryWithGradientContainer(
-                                    testVideos: testVideos),
-                              ),
-                              const SizedBox(height: 24),
-                              SizedBox(
-                                height: isMobile ? 350 : 220,
-                                child: ImageGallery(
-                                  images: ProjectMediaViewer.testImages,
-                                  crossAxisCount: gridCrossAxisCount,
-                                ),
-                              ),
-                            ],
-                          ),
-                  ),
+                SliverToBoxAdapter(
+                  child: (isBigDesktop || isDesktop || isTablet)
+                      ? ProjectMediaDesktopLayout(
+                          name: name,
+                          description: description,
+                          images: testImages,
+                          videos: testVideos,
+                          isTablet: isTablet,
+                          horizontalPadding: isBigDesktop ? width * .07 : 24,
+                          verticalPadding: isBigDesktop ? 36 : 16,
+                        )
+                      : ProjectMediaMobileLayout(
+                          description: description,
+                          images: testImages,
+                          videos: testVideos,
+                          gridCrossAxisCount: gridCrossAxisCount,
+                          isMobile: width < 700,
+                          horizontalPadding: isBigMobile ? 48 : 24,
+                          verticalPadding: isBigMobile ? 36 : 16,
+                        ),
                 )
               ],
             );
           },
         ),
       ),
-    );
-  }
-}
-
-class VideoGalleryWithGradientContainer extends StatelessWidget {
-  const VideoGalleryWithGradientContainer({
-    super.key,
-    required this.testVideos,
-  });
-
-  final List<String> testVideos;
-
-  @override
-  Widget build(BuildContext context) {
-    return AspectRatio(
-      aspectRatio: 2 / 1,
-      child: Container(
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(24),
-          gradient: const LinearGradient(
-            colors: [Colors.pinkAccent, Colors.blue],
-            begin: Alignment.topLeft,
-            end: Alignment.bottomRight,
-          ),
-        ),
-        padding: const EdgeInsets.all(4),
-        child: Container(
-          clipBehavior: Clip.hardEdge,
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(20),
-            color: Colors.black,
-          ),
-          child: VideoGallery(
-            videos: ProjectMediaViewer.testVideos,
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ImageGallery extends StatelessWidget {
-  final List<String> images;
-  final int crossAxisCount;
-  const ImageGallery({required this.images, this.crossAxisCount = 3});
-
-  @override
-  Widget build(BuildContext context) {
-    return GridView.builder(
-      gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisCount: crossAxisCount,
-        mainAxisSpacing: 16,
-        crossAxisSpacing: 16,
-        childAspectRatio: 1,
-      ),
-      itemCount: images.length,
-      itemBuilder: (context, index) {
-        return HoverableImageContainer(
-          imageUrl: images[index],
-          onTap: () {
-            showDialog(
-              context: context,
-              builder: (_) => ImageGalleryDialog(
-                images: images,
-                initialIndex: index,
-              ),
-            );
-          },
-          height: 120,
-          width: 120,
-        );
-      },
-      shrinkWrap: true,
-      physics: const NeverScrollableScrollPhysics(),
-    );
-  }
-}
-
-class ImageGalleryDialog extends StatefulWidget {
-  final List<String> images;
-  final int initialIndex;
-  const ImageGalleryDialog({required this.images, required this.initialIndex});
-  @override
-  State<ImageGalleryDialog> createState() => _ImageGalleryDialogState();
-}
-
-class _ImageGalleryDialogState extends State<ImageGalleryDialog> {
-  late PageController _controller;
-  late int _currentIndex;
-
-  @override
-  void initState() {
-    super.initState();
-    _currentIndex = widget.initialIndex;
-    _controller = PageController(initialPage: _currentIndex);
-  }
-
-  void _goTo(int newIndex) {
-    if (newIndex >= 0 && newIndex < widget.images.length) {
-      setState(() => _currentIndex = newIndex);
-      _controller.jumpToPage(newIndex);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      backgroundColor: Colors.black,
-      insetPadding: const EdgeInsets.all(16),
-      child: Stack(
-        alignment: Alignment.center,
-        children: [
-          PageView.builder(
-            controller: _controller,
-            itemCount: widget.images.length,
-            onPageChanged: (i) => setState(() => _currentIndex = i),
-            itemBuilder: (context, index) {
-              return PhotoView(
-                imageProvider: NetworkImage(widget.images[index]),
-                backgroundDecoration: const BoxDecoration(color: Colors.black),
-              );
-            },
-          ),
-          Positioned(
-            left: 16,
-            child: IconButton(
-              icon: const Icon(Icons.arrow_back_ios, color: Colors.white),
-              onPressed:
-                  _currentIndex > 0 ? () => _goTo(_currentIndex - 1) : null,
-            ),
-          ),
-          Positioned(
-            right: 16,
-            child: IconButton(
-              icon: const Icon(Icons.arrow_forward_ios, color: Colors.white),
-              onPressed: _currentIndex < widget.images.length - 1
-                  ? () => _goTo(_currentIndex + 1)
-                  : null,
-            ),
-          ),
-          Positioned(
-            top: 16,
-            right: 16,
-            child: IconButton(
-              icon: const Icon(Icons.close, color: Colors.white),
-              onPressed: () => Navigator.of(context).pop(),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class VideoGallery extends StatelessWidget {
-  final List<String> videos;
-  const VideoGallery({required this.videos});
-  @override
-  Widget build(BuildContext context) {
-    return PageView.builder(
-      itemCount: videos.length,
-      itemBuilder: (context, index) {
-        return _VideoPlayerWidget(videoPath: videos[index]);
-      },
-    );
-  }
-}
-
-class _VideoPlayerWidget extends StatefulWidget {
-  final String videoPath;
-  const _VideoPlayerWidget({required this.videoPath});
-  @override
-  State<_VideoPlayerWidget> createState() => _VideoPlayerWidgetState();
-}
-
-class _VideoPlayerWidgetState extends State<_VideoPlayerWidget> {
-  late VideoPlayerController _controller;
-  ChewieController? _chewieController;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = VideoPlayerController.network(widget.videoPath)
-      ..initialize().then((_) {
-        setState(() {
-          _chewieController = ChewieController(
-            videoPlayerController: _controller,
-            autoPlay: false,
-            looping: false,
-            showControls: true,
-            allowFullScreen: true,
-            allowMuting: true,
-          );
-        });
-      });
-  }
-
-  @override
-  void dispose() {
-    _chewieController?.dispose();
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: _chewieController != null && _controller.value.isInitialized
-          ? Chewie(controller: _chewieController!)
-          : const CircularProgressIndicator(),
     );
   }
 }
