+++ lib/view/projects/components/project_media_widgets.dart
@@ -1,0 +1,209 @@
+import 'package:flutter/material.dart';
+import 'hoverable_image_container.dart';
+import 'package:photo_view/photo_view.dart';
+import 'package:video_player/video_player.dart';
+import 'package:chewie/chewie.dart';
+
+class VideoGalleryWithGradientContainer extends StatelessWidget {
+  final List<String> videos;
+  const VideoGalleryWithGradientContainer({required this.videos, Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return AspectRatio(
+      aspectRatio: 2 / 1,
+      child: Container(
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(24),
+          gradient: const LinearGradient(
+            colors: [Colors.pinkAccent, Colors.blue],
+            begin: Alignment.topLeft,
+            end: Alignment.bottomRight,
+          ),
+        ),
+        padding: const EdgeInsets.all(4),
+        child: Container(
+          clipBehavior: Clip.hardEdge,
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(20),
+            color: Colors.black,
+          ),
+          child: VideoGallery(videos: videos),
+        ),
+      ),
+    );
+  }
+}
+
+class ImageGallery extends StatelessWidget {
+  final List<String> images;
+  final int crossAxisCount;
+  const ImageGallery({required this.images, this.crossAxisCount = 3, Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return GridView.builder(
+      gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+        crossAxisCount: crossAxisCount,
+        mainAxisSpacing: 16,
+        crossAxisSpacing: 16,
+        childAspectRatio: 1,
+      ),
+      itemCount: images.length,
+      itemBuilder: (context, index) {
+        return HoverableImageContainer(
+          imageUrl: images[index],
+          onTap: () {
+            showDialog(
+              context: context,
+              builder: (_) => ImageGalleryDialog(
+                images: images,
+                initialIndex: index,
+              ),
+            );
+          },
+          height: 120,
+          width: 120,
+        );
+      },
+      shrinkWrap: true,
+      physics: const NeverScrollableScrollPhysics(),
+    );
+  }
+}
+
+class ImageGalleryDialog extends StatefulWidget {
+  final List<String> images;
+  final int initialIndex;
+  const ImageGalleryDialog({required this.images, required this.initialIndex, Key? key}) : super(key: key);
+  @override
+  State<ImageGalleryDialog> createState() => _ImageGalleryDialogState();
+}
+
+class _ImageGalleryDialogState extends State<ImageGalleryDialog> {
+  late PageController _controller;
+  late int _currentIndex;
+
+  @override
+  void initState() {
+    super.initState();
+    _currentIndex = widget.initialIndex;
+    _controller = PageController(initialPage: _currentIndex);
+  }
+
+  void _goTo(int newIndex) {
+    if (newIndex >= 0 && newIndex < widget.images.length) {
+      setState(() => _currentIndex = newIndex);
+      _controller.jumpToPage(newIndex);
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Dialog(
+      backgroundColor: Colors.black,
+      insetPadding: const EdgeInsets.all(16),
+      child: Stack(
+        alignment: Alignment.center,
+        children: [
+          PageView.builder(
+            controller: _controller,
+            itemCount: widget.images.length,
+            onPageChanged: (i) => setState(() => _currentIndex = i),
+            itemBuilder: (context, index) {
+              return PhotoView(
+                imageProvider: NetworkImage(widget.images[index]),
+                backgroundDecoration: const BoxDecoration(color: Colors.black),
+              );
+            },
+          ),
+          Positioned(
+            left: 16,
+            child: IconButton(
+              icon: const Icon(Icons.arrow_back_ios, color: Colors.white),
+              onPressed:
+                  _currentIndex > 0 ? () => _goTo(_currentIndex - 1) : null,
+            ),
+          ),
+          Positioned(
+            right: 16,
+            child: IconButton(
+              icon: const Icon(Icons.arrow_forward_ios, color: Colors.white),
+              onPressed: _currentIndex < widget.images.length - 1
+                  ? () => _goTo(_currentIndex + 1)
+                  : null,
+            ),
+          ),
+          Positioned(
+            top: 16,
+            right: 16,
+            child: IconButton(
+              icon: const Icon(Icons.close, color: Colors.white),
+              onPressed: () => Navigator.of(context).pop(),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class VideoGallery extends StatelessWidget {
+  final List<String> videos;
+  const VideoGallery({required this.videos, Key? key}) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    return PageView.builder(
+      itemCount: videos.length,
+      itemBuilder: (context, index) {
+        return _VideoPlayerWidget(videoPath: videos[index]);
+      },
+    );
+  }
+}
+
+class _VideoPlayerWidget extends StatefulWidget {
+  final String videoPath;
+  const _VideoPlayerWidget({required this.videoPath, Key? key}) : super(key: key);
+  @override
+  State<_VideoPlayerWidget> createState() => _VideoPlayerWidgetState();
+}
+
+class _VideoPlayerWidgetState extends State<_VideoPlayerWidget> {
+  late VideoPlayerController _controller;
+  ChewieController? _chewieController;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = VideoPlayerController.network(widget.videoPath)
+      ..initialize().then((_) {
+        setState(() {
+          _chewieController = ChewieController(
+            videoPlayerController: _controller,
+            autoPlay: false,
+            looping: false,
+            showControls: true,
+            allowFullScreen: true,
+            allowMuting: true,
+          );
+        });
+      });
+  }
+
+  @override
+  void dispose() {
+    _chewieController?.dispose();
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: _chewieController != null && _controller.value.isInitialized
+          ? Chewie(controller: _chewieController!)
+          : const CircularProgressIndicator(),
+    );
+  }
+}
